FINERACT-2181: Fix savings integration tests

+++ integration-tests/src/test/java/org/apache/fineract/integrationtests/savings/base/BaseSavingsIntegrationTest.java
@@ -129,8 +129,8 @@ public class BaseSavingsIntegrationTest extends IntegrationTest {
     }
 
     protected PostSavingsAccountsRequest applySavingsRequest(Long clientId, Long productId, String submittedDate) {
-        return new PostSavingsAccountsRequest().clientId(clientId.intValue()).productId(productId.intValue()).dateFormat(DATETIME_PATTERN)
-                .locale("en").submittedOnDate(submittedDate);
+        return new PostSavingsAccountsRequest().clientId(clientId).productId(productId).dateFormat(DATETIME_PATTERN).locale("en")
+                .submittedOnDate(submittedDate);
     }
 
     protected PostSavingsAccountsResponse applySavingsAccount(PostSavingsAccountsRequest request) {
